backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.0.1
beanie==2.0.0
certifi==2026.4.22
click==8.3.2
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
lazy-model==0.3.0
motor==3.7.1
packaging==26.2
passlib==1.7.4
pluggy==1.6.0
pydantic==2.13.2
pydantic-settings==2.13.1
pydantic_core==2.46.2
Pygments==2.20.0
PyJWT==2.12.1
pymongo==4.16.0
pytest==9.0.3
python-dotenv==1.2.2
python-multipart==0.0.26
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
